fix: trace exporter
